piecestore: actually go up to the max allocation step
in poking around and adding instrumentation over the weekend,
i discovered that most orders are not actually using the
"next allocation step" size. in fact, because the piecestore
Upload type is an io.Writer, it's tough to control the amount
of data that gets put into an order.

this change converts the internal details of the Uploader to
use io.Readers, which allows the increasing order size logic
control exactly how much goes into an order, instead of hoping
for higher level buffers and flushes to do it correctly.

this change also removes the concept of a per-storage-node
upload buffer limit. the piecestore client must have enough
memory to store the max allocation step size, so just use that
instead.

Change-Id: I617ba786c8b2c6d3574a19d97a1b13de7e90aee5