all: Enable gosec linter
Enable gosec linter.

Change the usage of an unsafe pointer by the specific type for avoiding
to disable in some line due to the linter error message:
"G103: Use of unsafe calls should be audited" and after discussed that
for now it has not been proved that using a mutex impacts the
performance considerably versu using and unsafe pointer with an atomic
operation.

Ticket: https://storjlabs.atlassian.net/browse/SM-432

Change-Id: I25dc37b3570efa67d8fd0c61775bed31b4d81395