project: Run telemetry during project lifecycle
In order to send uplink telemetry to our statreceiver, we instantiate a
telemetry client when we open an uplink project. When we close the
project, we flush telemetry data.

Telemetry reporting is opt-in, with a simple function that attaches
default information to the context, and a separate function to add
custom options to the context. Telemetry is only sent if enabled in the
ctx.

Change-Id: I243d3de59cb6f147453b52b903d365783c61b7e1