Fix issues with sharing access with share prefix
This change adopts the fix from storj.io/common for sharing whole
bucket.

It also fixes removing the trailing slash before encrypting the share
prefix and list prefix, which caused a number of other issue with
sharing accesss.

Test cases are added for downloading and listing objects with share
prefix.

Change-Id: I7327491e2313e19d6843cb770acd37d422c968cd