upload: close readers on Abort
Resources can be leaked in the case where you create an
upload and call Abort on it because nothing cancels the
blocked reads on the io.Pipe created in the stream upload.

This change adds an Abort method to the stream upload and
calls it appropriately. That abort method closes the io.Pipe
with an error so that reads become unblocked.

Change-Id: I06b91d624603971163c27d3d341135662593cf73