Merge 'master' branch (#18)
* private/metainfo: simplify fetching object metadata

This change removes unnecessary code and simplify code around fetching
object metadata.

Change-Id: I645c5abeb259dfa325920a9489a4d49f2bf50bc8

* private/storage/streams: send encryption parameters with begin object
request

We are not sending those parameters with request as with current
implementation we are putting them into StreamMeta but it will be needed
after metainfo refactoring.

At the moment its hard to test that this value is set correctly as its
not used on satellite side too.

Change-Id: Ieca3a7c689b7c5d91d64502300d8a5bba9e1b6dd

* testsuite: ensure we print full info on failure

Change-Id: Idff5d60d395027461b89751f7dfd7dc0104e21e8

* private/metainfo: remove unused MutableObject interface

Change-Id: Id60bcf028536237d98d384142507392a651c0731

* private/metainfo: remove MutableStream interface

Currently the dynamic metadata required by upload is still awkward.
We'll need to find a better abstraction that would make mutable stream
easier to use, possibly, getting rid of it altogether.

Change-Id: I573321aa88044c679138b4db5cfa4575a9e93438

* private/metainfo: remove unused ReadOnlyStream interface

Change-Id: Iba917ddd49e0d739024fe27851b1a59a99025986

* private/storage/streams: remove unused interface

Change-Id: Ifc49e4c05eab1d550d2e4504151ce76cada5ed6c

* private/storage/streams: preserve threshold err

Change-Id: Idefff8f8a762f4b33632e7ecdb26b16da6202e89

* private/storage/streams: handle potentially wrapped io.EOF

We should not wrap io.EOF, however it's better to be safer when checking
for io.EOF.

Change-Id: I11bad5bd659de2ef8361806eca201b864c9f1f7a

* private/access2: create access2 package to maintain uplink api while exposing new methods

This change is needed for webAssembly changes.

Change-Id: I43613617b5280dded501e500de08eec12dc8c8eb

* private/metainfo: remove unused fields

Change-Id: Iaeecd328da562b72299df0d1ecae8b37a3fc4e2a

* testsuite: skip flaky test

Change-Id: Ife07593e2a2ec8d0a86e3d39be76f5fd9b2deba3

* mod: bump drpc@v0.0.16

Latest DRPC has a fix for concurrent method invocations.

Bumping libuplink size due to quic adding ~0.5MB to the binary.

Change-Id: I8debfb37d5e20ed61359a8ddc1adb9fae267f466

* ci: check also ./testsuite

Go pattern matching only matches the current module. This means that
submodules need separate checks.

Also includes linter fixes.

Change-Id: I49c383eb77e0c416d59c44d4732a3d3156841e1a

* ci: ensure cockroach doesn't pollute repo

Cockroach 20.2 started automatically profiling, this is a workaround to
disable it and ensure it doesn't create any folders and files in the
repository.

Change-Id: I9208239862dbf074ae13b957daed658f1401d69f

* testsuite/scripts: stop using Minio fork for gateway installation

Change-Id: I8e8cd2584042b208b1fb4204f60ef00970dcd6fd

* private/stream: use storj.Object instead of ReadOnlyStream

Change-Id: Ied87e5756f13b2f694485ebcd5781fe049280673

* private/metainfo: remove GetObjectStream

Change-Id: I390e044e6ce40be0b846769096e6dbc0244dadcd

* private/metainfo: remove db reference from MutableObject

Change-Id: Ib6bb1fd3624b70ba6438f480156cb789a29eacab

* private: remove usage of DeprecatedNumberOfSegments

All old-style objects were migrated on all production satellites to have
the number of segments in the unencrypted metadata. The
DeprecatedNumberofSegments field is not necessary anymore.

Change-Id: Ia771c3c063dd4db6e67bfd5492ec3edaa637a685

* private/metainfo: make custom types for GetObject and ListObjects
results

With this change:
* storj.ObjectInfo was changed to type RawObjectItem
* storj.ObjectListItem was changed to type RawObjectListItem
* code simplification around metainfo.DB methods

Change-Id: Ia70a4e1446d98e6694870c82cac15931445c7b78

* uplink: dial every time and use pooled connections

This fixes a problem where a project could end up forever
dead if the underlying connection it dialed when opening
ended up dead. This was being observed in the gateway.

Change-Id: Ica25322fc6c52936df82ecad9038e1d1c02359bd

* uplink: reorganize parameters initialization

This change contains two cleanups:
* Move hardcoded values initialization to OpenProject method to avoid
redoing this work on every dial
* Move calculation of maxEncryptedSegmentSize to streams.Store Put
method to stop doing it for every dial

Change-Id: Idc7af00afa35936d89ee8f1bb07824d64f41715b

* Access Satellite Address

Add `(*Access).SatelliteAddress` which returns the satellite node url.

Change-Id: I74f958843d0184139bc6e1b979d51a91984c3160

* uplink: clean up some pooled connection changes

Change-Id: I85182107f41775a8f708d2d2f4efd811f3f72c2d

Co-authored-by: Egon Elbre <egonelbre@gmail.com>
Co-authored-by: jessicagreben <Jessica.greben1+github@gmail.com>
Co-authored-by: Jeff Wendling <leterip@gmail.com>
Co-authored-by: Kaloyan Raev <kaloyan@storj.io>
Co-authored-by: Caleb Case <caleb@storj.io>