uplink: always use resolved satellite address
we have a list of known addresses that we add the id to hard
coded. when provided one of these addresses, we must always
add the id in order to maintain backward compatibility. in order
to make it so that this is not forgotten, change the apis to
use parsed storj.NodeURLs when possible, and change the
access constructors to parse and check for known addresses.

Change-Id: I0a3180874aa311483b12fe535c274357baeda857