private/storage/streams: do not download last segment prematurely
https://storjlabs.atlassian.net/browse/SM-1922

If we download the last segment first from the satellite, its order
limits are created at the very beginning of the download. If the
complete file download takes over 1 hour, by the time the last segment
pieces are downloaded, the order limits already expired. The storage
nodes will reject to provide the pieces, and the download will fail for
the last segment.

Change-Id: I04f5e05a04de872ed3c0163a59b1b303aad649a3