uplink: stop restarting traces
we were resetting/restarting traces because of long-lived
contexts, but resetting/restarting traces are too painful
(see https://storjlabs.atlassian.net/wiki/spaces/ENG/pages/1442742314/Tracing+client+operations
for an example).
we should instead add trace restarting in the specific
locations where it matters, and not for everything that
uses the uplink.

Change-Id: I914bd407f36fb0005c5e2f8201a188fdce5ccb38