testsuite/private/multipart: fix flaky TestListParts
The test does a multipart uploads of an object with random size between
100K and 600K in 2 parts. The max segment size is configured to 50K.
This makes between 3 and 13 segments for the objects.

ListObjectParts is invoked with maxParts = 10, but maxParts seems to
limit the number of segments returned, not parts. Thus, sometimes this
is not enough to return all segments in a single ListSegments call, so
the expected part size does not match what has been calculated by the
uplink.

As a temporary workaround, until we fix the proper handling of maxParts,
this PR increasing maxParts to 100.

Change-Id: Ic2545f180ccf719a8a29c6b617ed49fad903210e