private/{metainfo,multipart}: commit and list part's ETag
PutObjectPart encrypts and commits the ETag provided by the client to
the satellite. ListObjectParts decrypts and returns the ETag.

The ETag is encrypted with a key derived from the randomly generated
segment's content key. This is to ensure that a unique (key, nonce) pair
is used for the ETag encryption.

Change-Id: I2287d77241f22950e7000c00d2051309b2315a00