uplink: make negative Offset in DownloadOptions download suffix
Currently there was no way to directly read the suffix of the content.
Reading the suffix is needed to support HTTP Range spec.

An alternative approach would be to introduce a new method, however
this would increase the API size. Changing negative offset to mean
suffix is easier and people probably do not rely on this behavior.

Change-Id: I94a5aa75f13101a067383fb7bc5a14e91418521a