private/metainfo: clean up+fix object listing special cases
Listing objects was not handled correctly for the key "/".
This is because the previous code would get the parent of the
key by removing a trailing "/" (which it ensured existed) and
encrypting the result. This works for every key except "/".

To see why, first, define the path component encryption function
to be enc, and the object key encryption function to be Enc.
In other words, Enc("foo/bar") == enc("foo")/enc("bar"). As a
small complication, we have the special case that Enc("") == ""
but enc("") != "". Consider the encryption of the key "/".
We have Enc("/") == enc("")/enc("") and so we should list the
prefix enc(""). But if we remove the trailing slash, instead of
getting the correct listing prefix, we get "", because of
the specail case, Enc("") == "".

As far as fixing this goes, we could try to just always encrypt
whatever key is passed in without modifying it and then remove
the last encrypted component, but then we have problems decrypting
the results of the list because the encryption key that should
be used is not available. Since encryption keys are derived from
unencrypted components, we would have to find the unencrypted
key that matches the prefix, but that doesn't always exist (there
is no unencrypted key that maps to enc("")/, for example).

So, we change to use the new encryption.GetPrefixInfo function
from the common module which does all the proper encryption
and returns the appropriate prefixes, keys, and other information
necessary to do listing correctly.

Finally, the ListPendingObjectStreams method used a lot of the
same code as the ListObjects method, but they are very different
in the sense that the latter takes a prefix whereas the former
takes a specific path. The encryption keys used to decrypt
the results from the server are different. Additionally, the
ListPendingObjectStreams method returns objects where the key
is the full encrypted path, but ListObjects just returns the
suffix. Because of these differences, some of the shared code
no longer is easily reused across the two methods.

Change-Id: I8075c8ad991bc39cb96c808c6f0be4a89831d9e3