private/storage/streams: delete cancelled upload
Previously we were leaving cancelled uploads to be cleaned up by zombie
cleanup, however, it's better to send a delete message to the satellite.
We weren't able to do this before, because we didn't have a way to
specify the stream id -- to ensure that the correct upload gets deleted.

Change-Id: I9ae235bc38d70393500625899852c9994dbb943c