private/metaclient: retry satellite requests
Add retry functionality to the uplink. Ultimately this should be moved
to the common repository, since we are using it in other places (e.g.
linksharing).

When certain errors are received involving satellite communication, we
will wait some time and retry. The wait time is increased each time
until the final attempt, which is capped at 3 seconds. If there is still
a failure at this point, the uplink shows an error.

Change-Id: Ib74b768999dbe2d38565a7adf8b9958dc9fec3ec