uplink: fix ETag encryption/decryption
This change fixes two things. First is how we decrypting
ETags because while listing parts we forgot to check only
last segment and we tried to decrypt empty ETag.

Second issue is related to uploading parts where plain size is
equal to maximum segment size (64MiB). In such case we are
skipping zero-byte segment at the end but in such case
skipped segment content key could be used to encrypt ETag
and that would make impossible to decrypt that ETag later.

Change-Id: I2cb8bbd7d7cec177cdd3c9f82f9f8d9aac9f8239