private/metaclient: stop using object RS
At the beginning, we were keeping only object redundancy scheme
values but at some point we decided that we maybe need to do
that per segment. We implemented that in a way to support both
approaches until all satellites will deploy needed code. It was
long time ago and we can remove obsolete code.

Change-Id: Ifef8b993d14a982deaa090793a6e4632cf09bb30