config.go: add maximumBufferSize for DRPC streams
This change adds the maximumBufferSize option to Config using linkname
hack. This is similar to the existing pool and connector, as we don't
want these options to be exposed by default. This change is primarily
needed to be able to configure the underlying DRPC connection manager to
drop large buffers from inside of, e.g., Gateway-MT or Link Sharing
Service.

Updates storj/gateway-mt#135

Change-Id: I4534308849b7233523e00bf57eabd02fac0a107a