private/piecestore: bump upload/download default timeout
for clients with low bandwidth, doing many uploads at once will
cause all of them to seem like they're taking a long time. if
the timeout is too short, then they can't actually upload. for
a step of 256 KiB, if you're doing 4 parallel uploads, each of
which makes 110 connections, you need a minimum upload bandwidth
of about 15Mbits, though, something seems off about this math
because we have reports of 2 parallel uploads failing with 25Mbits.

these timeouts are only necessary for cleaning up essentially
dead tcp connections, and only need to kill them before resource
exhaustion hits, which tends to take hours for our production
services. for a default, we really should be as conservative
possible so that low bandwidth situations still work and can
tune to be more aggressive when the situation allows and is
necessary.

in general, i'd prefer if these timeouts were handled by the
kernel using tcp keep alives or something, but for whatever
reason, i've been unable to make that reliable over the last
decade or so. lol.

Change-Id: I04544b51692839d79ede9ebce8bbdb7848589025