uplink: add CopyObject method
This method is an addition to uplink public API. It provides
functionality to copy object to different  bucket and/or key.
It combines two requests to satellite but from user perspective
its just single operation.

Method gets metadata and segments encrypted keys and nonces then
it decrypts keys used for encryption and encrypts them using
new object bucket and key. Result of this operation is send to
satellite.

Resolves https://github.com/storj/uplink/issues/79

Change-Id: I99e7d17d7782d26c9bcafd2400e61c10adb06f12