private/piecestore: wrap closing errors with CloseError
Some callers of the piecestore code (in particular, the repairer; see
https://github.com/storj/storj/issues/4687) need to be able to
differentiate between "stream errors encountered during Dial" and
"stream errors encountered during Close".

This is probably not the most beautiful way to accomplish that, but it
should at least be backwards compatible and it gets the job done.

I'm very open to suggestions about how to do this better.

Change-Id: Ia574adb035af03ceeeed25188a1bcfd3fd2834b8