edge: make authservice protocol configurable
Before switching to mandatory DRPC protocol for authservice, it was possible to use `uplink share` in a dev environment.

Unfortunatelly this feature is broken in `storj-up` as we require to set up TLS.

This chance makes it possible to call authservice in an insecure way.

Should be safe, as it works only if authservice also supports it, and production servers doesn't.

Change-Id: I33044f7a845673ae7319ff8349edc2ad6737f0f0