uplink: report path checksum when metrics are opted in
we are trying to analyze network performance. it would be
nice to be able to cluster operations on specific objects
within these metrics, but we obviously don't want to identify
paths or any user identifiable data.

for users who have explicitly opted in to data analytics
(e.g., enabled eventkit), this change adds reporting of
a "path checksum", which is defined as the first 16
bytes of an hmac of the encrypted path with the
phrase "event".

it's worth pointing out that this reporting is strictly less
data than a satellite operator already has.

Change-Id: I4de40f029a05a8aca90cd6f2895de18f7ca6aa8f