testsuite: use current uplink for tests
For a moment we stopped using current uplink with testsuite to be
able to do refactoring which involved storj/storj and storj/uplink.
Now we can bring back link to current uplink code in testsuite.

As an addition testsuite dependencies were updated and that requires
change in test to reflect "metainfo.multiple-versions" flag enabled
by default now for satellites.

Change-Id: I7dbc08a4c85ce9684969abc33786ce10dd4b557e