uplink/piecestore: send order as part of first message
the reason for this change is to get the order into the
noise handshake packet.

unfortunately, doing this indiscriminately is not
backwards compatible with old nodes. however, new nodes
(and all noise-supporting nodes) have this change:
https://review.dev.storj.io/c/storj/storj/+/9245
so, fortunately, we can combine the messages in the
cases that matter (noise support).

Change-Id: I787c44eb822060f2d28e90caf85952aff3f3cc1f