private/storage/streams/streambatcher: introduce Batcher
The stream batcher wraps the metainfo batch RPC for batch items on
behalf of a stream upload. It learns and injects the stream ID and
gathers information from batch items being issued.

This is a component of the new upload codepath and simplifies state
tracking and information gathering needed to commit the part/object.

Co-authored-by: Jeff Wendling <jeff@storj.io>

Change-Id: Ieeecc2d4ab8743a6c9b69babb7691554fd05da8b