private/storage/streams: segmenttracker package
The segment tracker is used to defer the scheduling of the batch item
that commits/creates the final segment in a multipart upload until all
segments have been scheduled and the eTag is available.

It encrypts the eTag using the last segment and injects it into the
CommitSegment/MakeInlineSegment request and then schedules the batch
item.

Fixes: https://github.com/storj/uplink/issues/134

Co-authored-by: Jeff Wendling <jeff@storj.io>

Change-Id: I79c3e460f4dbbbcf0012cec9017aeb794412f43f