private/storage/streams: pieceupload package
Provides the piece upload Manager, which tracks the pieces that need to
be uploaded as well as the results. The manager also is in charge of
obtaining new piece limits for failed piece uploads so the upload can be
retried.

The package also contains a function, UploadOne, which uploads a single
piece until successful. It uses the manager to determine the next piece
to upload.

Fixes: https://github.com/storj/uplink/issues/131
Fixes: https://github.com/storj/uplink/issues/135

Co-authored-by: Jeff Wendling <jeff@storj.io>

Change-Id: I8c36b0a9a9ca3da8cad4590676cf24508456a475