uplink: avoid nilnil
Returning nil, nil can easily introduce bugs related to
using nil pointer usage. Currently this will happen when
the user has permissions to delete a bucket, however
they do not have permissions to view it. We can still
return the bucket name, because the user already knew
it, so we are not leaking any information that wasn't
known.

Change-Id: Id29d628098925956f1511d94f104f58ff01cd3b8