private/piecestore: re-enable noise
noise was disabled due to finding issues that are now fixed:

 * https://review.dev.storj.io/c/storj/uplink/+/9620
 * https://review.dev.storj.io/c/storj/drpc/+/9713

The drpc fix should probably get a bit more explanation -
without it, long tail cancelation was broken, because
canceling the context on a drpc stream did not interrupt a
stalled close notification. with it, canceling the context
now does interrupt a stalled close notification, which fixes
long tail cancelation.

Change-Id: I973e90f2c6f3b26f7a5c686f216af9bd4cbace36