private/storage/streams: segmentupload package
This new package is in charge of uploading a single segment. It takes
the BeginSegment response as input and returns a CommitSegment request
as output.

It uploads pieces concurrently, capping the number of concurrent uploads
to the optimal threshold of the redundancy strategy plus a (currently
constant) safety margin.

It handles long tails by cancelling outstanding piece uploads as soon as
the optimal threshold has been reached.

Change-Id: I93d45dd2ec726f80fb4cf70dae0662ab3f091a2a