Revert "internal/private: use different connection pool for satellite/storagenode connections"
This reverts commit c1744bf560eeb6b73889aca9f63c5e6d5e214335.

Reason for revert:

After I started to use this commit, I realized that I am unhappy with it.

It works well when you would like to use two different pools.

But gateway-mt supposed to use two persistent pool. Using it with the factory pattern is not easy and confusing as it looks like we delegates the responsibility of opening/closing the connection. But it should be delegated.

Change-Id: Iea87d6a15f90b2ba62758f73327053a07417d117