config: deprecate dialing options and remove transport.SetConnector
We are starting to add different dialing options based on what
nodes support. We already have TCP vs QUIC for the underlying
transport, but we're considering adding TCP_FASTOPEN and
potentially other transport strategies, based on what the node
supports. Since this is potentially dependent on the node, we
need the library able to make these dialing decisions.

If the config allows an unnuanced overriding of dialing
behavior, then a lot of options are off the table. Let's
start to claw back ownership of these decisions.

Change-Id: I40653b35bfefb886f2be0553c77cf3c58b75c6bf