private/piecestore: bump default allocation steps
When we download a piece, we send incremetal orders to storagenode.

As of now we start from 64k, and multiply the allocation step with 1.5 every time up to the max step of 256k.

To download a 2.2MByte piece (64Mbyte segment size / 29 RS parameter) we need ~11 RPC requests and responses.
(Orders will be sent as requests, with these max sizes: 64k, 160k, 304k, 520k, 776k, 1032k, 1288k, 1544k, 1800k, 2056k, 2312k).

We need to send incremental orders to motivate storagenodes to serve the request (in the hope of more requests), also to decrease the read amplification of using 39 parallel downloads.

Long-term this can be avoided with:
 * Choosing the fastest 29 with measuring the speed of connection (assuming we do have at least one handshake ;-).
 * And/or maintain reputation stat for each storagenode.

Until that, we can bump these numbers to slightly higher numbers. For example using 256k (with 550k max), we limit the RPC requests to 5 (instead of 11).

Measurements showed, that small file downloads (like 4MB) can show ~100 ms improvement with this change.

550k max is chosen to have 64MByte/29 = 2.2MByte in the 5th step.

It doesn't change the max amplification (customer may pay 39/29 = 134% egress, if storagenodes have very similar speed, which is the case in EU). But they may have small difference with the long-tail cancelled downloads.
User may prepay (550-256)*(39-29) = 294k more bytes in very unlucky cases. (but max of the amplification is (39-29) * (64MB/29) = 22.5 Mbyte (this is the full 34%).

Change-Id: Id9054fd66091240151ea9a2139f1cd2ecb4464e9