pieceupload: limit number of retries
this change adds a test that uploads eventually fail
when there are no nodes to upload to, and changes the
piecemanager to go into a failure state if any exchange
pieces calls fail. the reason it goes into a permanent
failure state is because it assumes the underlying client
that issues the api request has sufficient retries, and
layering retries leads to exponential growth.

Change-Id: Icd759ee903a1c80346e0dcf1d7c228952f2d8921