storage/streams: finish the splitter with original error
uplink/clients/servers are very sensitive to specific errors
for proper diagnosis and display of the error to the user.
for example, specific status codes should be presented in the
s3 compativble API for certain errors. finish the splitter
with as much of an untouched error as possible to make this
analysis more robust.

Change-Id: I1832e19575d3e5a290d5ad95195c1fd7ff5baa4a