private/storage/streams/buffer: fix memory leak
the upload code path is careful to close the underlying
buffers and backends when it is done reading and writing
them, but the memory backend didn't actually free the
resources when closed. combined with the fact that we keep
a reference to all of the segments for the duration of the
upload, this meant that we were storing the entire upload
data in memory until the upload was complete. oops.

Change-Id: I3334a06ddd0346b09fc3bcd8870ccbb83b6ea108