private/storage/streams/streamupload: fix racy test
a test would start 3 segments: 1 good, 1 bad, and 1
that blocks forever. the bad segment would return
an error, which caused the context to be canceled.
under certain schedulings, the segment that blocks
forever would return an error before the bad segment
and so "context canceled" would be the recorded
error rather than the "wait failed" erorr from the
bad segment, causing the test to fail.

really, we don't want to just return out "context
canceled" if some other error happened, so this
introduces the cancelGroup type that is like an
errorgroup but controls the context cancel itself
and only cancels after it has recorded the error.

no more flakes, more reliable errors.

Change-Id: I47f002faee952287845441b22cc8e90257d06c9f