private/storage/streams: performance improvements
under very high bandwidth situations (10GBit+), the speed
of the upload is CPU limited and bottlenecked on locks.
this adds some fast path escapes that avoid locks for the
cursor and removes locks on the memory backend because
the cursor serializes access to it already. it does need
to use atomics to update the length, however.

additionally, the streamuploader will issue requests to
the satellite in parallel if there's buffering from the
stream which allows better pipelining of segment uploads.

these changes gave a >30% speedup on a 100Gbit machine
which is now seemingly bounded by memory bandwidth.

Change-Id: I3dfe47798a8b11f5434b9eec97519e9148a98e2b