private/storage/streams/streamupload: fix cancelgroup handling
If a segment fails to upload, the cancelgroup goroutine returns an error
, which cancels the cancelgroup context. This context is used by the
segment source as well, which causes calls to Next() to return
context.Canceled.

We really want the error from the cancelgroup though. This patch changes
the code to use the cancelgroup error unless the upload context itself
is done.

Change-Id: I3704e8be217a408bb13c19827d940721a3eaf0bc